Rename `RawDataPanePicker` to `RawDataPicker`
