--- frontend/src/metabase/containers/DataPicker/RawDataPanePicker/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./RawDataPanePickerContainer";
