+++ frontend/src/metabase/containers/DataPicker/RawDataPicker/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./RawDataPickerContainer";
